refactoring and adding some styling
